- Ajouté le support pour la sérialisation de string[] dans Common.Text. - Complété Common.Types.Comparer pour comparer les tableaux soi-même; ceci permet de déterminer si deux tableaux ont un contenu identique. - Corrigé une erreur dans Text.TextConverter. En cas de substitutions, la copie du texte résultant ne se faisait pas correctement. - Adapté CheckUnicode à la nouvelle politique (considérer \r comme illégal et remplacer \n par un saut de paragraphe). - Adapté d'autres tests de Text.Tests qui ne passaient plus à la suite de modifications et de corrections. - Ajouté un champ 'Features' à Text.Properties.FontProperty. - Ajouté le support pour le crénage dans OpenType.Font.
git-svn-id: https://svn.opac.ch/svn/cr@3170 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Common.OpenType/Tables.cs
@@ -1438,7 +1438,7 @@ namespace Epsitec.Common.OpenType
 		
 		public int GetKernValue(int n)
 		{
-			return this.ReadInt16 (18+n*6);
+			return (short) this.ReadInt16 (18+n*6);
 		}
 		
 		public bool FindKernValue(int left, int right, out int value)
